finishing touches

+++ Assets/Scripts/CommandListManager.cs
@@ -5,36 +5,45 @@ using UnityEngine;
 public class CommandListManager : MonoBehaviour
 {
 	private List<MoveObjectCommand> commandList;
-	private int currentIndex = -1;
+	private int currentIndex;
 
 	private void Start()
 	{
 		commandList = new List<MoveObjectCommand>();
+		currentIndex = -1;
 	}
 
 	private void Update()
 	{
-		if (currentIndex > -1)
+		if (currentIndex > -1 && currentIndex < commandList.Count)
 		{
 			if (Input.GetKeyDown(KeyCode.Q))
 			{
 				commandList[currentIndex].Undo();
-				if (commandList[currentIndex].bIsCrate)
+				commandList[currentIndex].gameObject.GetComponent<AudioSource>().Play();
+				if (currentIndex - 1 > -1 && commandList[currentIndex - 1].bIsCrate)
 				{
 					commandList[currentIndex - 1].Undo();
+					commandList[currentIndex].gameObject.GetComponent<AudioSource>().Play();
 					currentIndex--;
 				}
 				currentIndex--;
 			}
 			else if (Input.GetKeyDown(KeyCode.E))
 			{
-				commandList[currentIndex].Execute();
-				if (currentIndex + 1 < commandList.Count && commandList[currentIndex + 1].bIsCrate)
+				if (currentIndex + 1 < commandList.Count)
 				{
-					commandList[currentIndex + 1].Execute();
 					currentIndex++;
+					commandList[currentIndex].Execute();
+					commandList[currentIndex].gameObject.GetComponent<AudioSource>().Play();
+
+					if (commandList[currentIndex].bIsCrate)
+					{
+						currentIndex++;
+						commandList[currentIndex].Execute();
+						commandList[currentIndex].gameObject.GetComponent<AudioSource>().Play();
+					}
 				}
-				currentIndex++;
 			}
 		}
 	}
@@ -43,9 +52,10 @@ public class CommandListManager : MonoBehaviour
 	{
 		currentIndex++;
 		MoveObjectCommand newCommand = new MoveObjectCommand(gameObject, newPosition);
-		commandList[currentIndex] = newCommand;
+		commandList.Insert(currentIndex, newCommand);
 
 		newCommand.Execute();
+		commandList[currentIndex].gameObject.GetComponent<AudioSource>().Play();
 		RemoveInvalidCommands();
 	}
 
@@ -54,7 +64,7 @@ public class CommandListManager : MonoBehaviour
 		if (currentIndex < 0 && commandList.Count > 0)
 			commandList.Clear();
 
-		if (commandList.Count > currentIndex + 1)
+		if (currentIndex < commandList.Count - 1 && currentIndex > -1)
 			commandList.RemoveRange(currentIndex + 1, commandList.Count - 1);
 	}
 }
